Util -> Embulk::Input::TwitterAds::Util
